refactor(rest): split converter.py into a converters/ package
The converter already dispatched through `ConverterRegistry` keyed on
(src_ext, target) with `(src_path, on_progress, **options)` handlers — a
different shape from the worker's job-kind handlers — so it keeps that
registry; the 3.7k-line module is split by handler family under
`ada.comms.rest.converters/` (registry, keys, ada_load, ada_export,
ada_pairs, pipelines, serializers, fea, mesh_step, step_stream), each moved
verbatim. `converters/__init__.py` fires the registrations in the original
order (decorated FEA deck pairs, then passthrough / trimesh / ada-loadable /
FEA result / glb->mesh / STEP stream exports, later ones overriding).
`converter.py` stays the facade: `convert()`, `result_bytes`,
`supported_extensions`, `TARGET_FORMATS`, `LEGACY_CONVERT_EXTS`, and every
previous name re-exported. Two tests patch the module where the patched
name is now looked up.
